feat: #27: Added Confirmation Dialog

+++ frontend/src/components/orders/OrderButtons.tsx
@@ -11,19 +11,37 @@ import HighlightOffIcon from '@mui/icons-material/HighlightOff';
 import DeleteIcon from '@mui/icons-material/Delete';
 import ScheduleSendIcon from '@mui/icons-material/ScheduleSend';
 import CakeIcon from '@mui/icons-material/Cake';
+import confirmDialog from "../../utils/confirmationDialog.tsx";
 
 export default function OrderButtons({order, onRefresh}: { order: Order, onRefresh: () => void, }) {
   const summary = useOrderPositionSummary(order);
   const navigate = useNavigate();
 
-  const handleRevoke = useCallback(() => {
-    api.orders.revokeOrder(order.id)
-      .then(() => onRefresh())
+  const handleRevoke = useCallback(async () => {
+    if (await confirmDialog({
+      title: 'Möchtest du die Bestellung wirklich widerrufen?',
+      content: <div style={{display: 'flex', flexDirection: 'column', gap: '1em'}}>
+        <Typography>
+          Die Bestellung wird hiermit widerrufen und in kurzer Zeit entfernt. Keiner kann seine Bestellung ändern oder
+          etwas hinzufügen!
+        </Typography>
+
+        <Typography variant="caption">
+          Das ist z.B. wenn man beim Bestellen merkt, dass das Restaurant geschlossen hat oder andere unerwartete
+          Probleme aufkommen.
+        </Typography>
+      </div>
+    })) {
+      api.orders.revokeOrder(order.id)
+        .then(() => onRefresh())
+    }
   }, [order.id, onRefresh]);
 
-  const handleDelete = useCallback(() => {
-    api.orders.deleteOrder(order.id)
-      .then(() => navigate({pathname: `/order`}, {replace: true}))
+  const handleDelete = useCallback(async () => {
+    if (await confirmDialog({title: 'Möchtest du die Bestellung wirklich löschen?'})) {
+      api.orders.deleteOrder(order.id)
+        .then(() => navigate({pathname: `/order`}, {replace: true}))
+    }
   }, [order.id, navigate]);
 
   const handleArchive = useCallback(() => {
@@ -31,26 +49,69 @@ export default function OrderButtons({order, onRefresh}: { order: Order, onRefre
       .then(() => onRefresh())
   }, [order.id, onRefresh]);
 
-  const handleOrdering = useCallback(() => {
-    api.orders.lockOrder(order.id)
-      .then(() => onRefresh())
+  const handleOrdering = useCallback(async () => {
+    if (await confirmDialog({
+      title: 'Möchtest du die Bestellung wirklich sperren?',
+      content:
+        <div style={{display: 'flex', flexDirection: 'column', gap: '1em'}}>
+          <Typography>
+            Die Bestellung wird hiermit gesperrt und keiner kann seine Bestellung ändern oder etwas hinzufügen!
+          </Typography>
+
+          <Typography variant="caption">
+            In der Zeit sollst du die Bestellung beim Restaurant aufgeben und danach hier bestätigen.
+            Machst du das nicht, wird die Bestellung in wenigen Minuten automatisch wieder entsperrt.
+          </Typography>
+        </div>
+    })) {
+      api.orders.lockOrder(order.id)
+        .then(() => onRefresh())
+    }
   }, [order.id, onRefresh]);
 
-  const handleOrderIsOrdered = useCallback(() => {
-    api.orders.orderIsNowOrdered(order.id)
-      .then(() => onRefresh())
+  const handleOrderIsOrdered = useCallback(async () => {
+    if (await confirmDialog({
+      title: 'ist die Bestellung erfolgreich bestellt worden?',
+      content:
+        <div style={{display: 'flex', flexDirection: 'column', gap: '1em'}}>
+          <Typography variant="caption">
+            Die Bestellung wurde erfolgreich bestellt. Änderungen ausser dem Bezahlen sind nicht mehr möglich.
+          </Typography>
+
+          <Typography color="error">
+            Vergiss nicht, das Eintreffen der Bestellung hier abzuhaken und dafür zu sorgen, dass alle bezahlt haben.
+          </Typography>
+        </div>
+    })) {
+      api.orders.orderIsNowOrdered(order.id)
+        .then(() => onRefresh())
+    }
   }, [order.id, onRefresh]);
 
-  const handleReopen = useCallback(() => {
-    api.orders.reopenOrder(order.id)
-      .then(() => onRefresh())
+  const handleReopen = useCallback(async () => {
+    if (await confirmDialog({
+      title: 'Möchtest du die Bestellung wieder entsperren?',
+      content:
+        <div style={{display: 'flex', flexDirection: 'column', gap: '1em'}}>
+          <Typography>
+            Die Bestellung wird wieder entsperrt und jeder kann seine Bestellung ändern oder etwas hinzufügen!
+          </Typography>
+
+          <Typography variant="caption">
+            Das soll benutzt werden, wenn z.B. das Restaurant noch nicht offen hat oder das Telefon besetzt ist und
+            später nochmal versucht werden muss zu bestellen.
+          </Typography>
+        </div>
+    })) {
+      api.orders.reopenOrder(order.id)
+        .then(() => onRefresh())
+    }
   }, [order.id, onRefresh]);
 
   const handleDelivery = useCallback(() => {
     api.orders.orderIsNowDelivered(order.id)
       .then(() => onRefresh())
   }, [order.id, onRefresh]);
-
 
   const revokeButton = <Button id="btn-order-state-revoke"
                                variant="contained"
